+++ packages/node/src/behaviors/window-covering/WindowCoveringServer.ts
@@ -73,11 +73,11 @@ const WC_PERCENT100THS_COEFFICIENT = 100;
  * In addition to Matter attributes, {@link WindowCoveringServerLogic.State} includes the following configuration
  * properties:
  *
- *   * supportsCalibration (default false): Set to true if the device supports calibration. You must implement
- *     {@link WindowCoveringServerLogic.executeCalibration} to perform actual calibration.
- *
  *   * supportsMaintenanceMode (default true): Set to false if the device has no maintenance mode
  *
+ * The internal state allows to configure implementation details when extending the class:
+ *   * supportsCalibration (default false): Set to true if the device supports calibration. You must implement
+ {@link WindowCoveringServerLogic.executeCalibration} to perform actual calibration.
  *   * disableOperationalModeHandling (default false): Set to true if you want to handle the operational status yourself
  *
  * When developing for specific hardware you should extend {@link WindowCoveringServer} and implement the following
@@ -147,7 +147,7 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
             this.state.targetPositionLiftPercent100ths = this.state.currentPositionLiftPercent100ths;
         }
 
-        if (!this.state.disableOperationalModeHandling) {
+        if (!this.internal.disableOperationalModeHandling) {
             // Keep position attributes (percentage and also absolute position) and operational state in sync
             this.reactTo(this.events.currentPositionLiftPercent100ths$Changing, this.#syncLiftCurrentPositions);
             this.reactTo(this.events.currentPositionTiltPercent100ths$Changing, this.#syncTiltCurrentPositions);
@@ -175,7 +175,7 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
         this.internal.inMaintenanceMode = !!mode.maintenanceMode;
 
         if (mode.calibrationMode) {
-            if (!this.state.supportsCalibration) {
+            if (!this.internal.supportsCalibration) {
                 throw new StatusResponseError("Calibration not supported", StatusCode.ConstraintError);
             }
             if (this.internal.calibrationMode === CalibrationMode.Running) {
@@ -188,7 +188,8 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
         }
 
         const configStatus = this.state.configStatus;
-        configStatus.operational = !mode.maintenanceMode || (mode.calibrationMode && !this.state.supportsCalibration);
+        configStatus.operational =
+            !mode.maintenanceMode || (mode.calibrationMode && !this.internal.supportsCalibration);
         configStatus.liftMovementReversed = !!mode.motorDirectionReversed;
         if (isDeepEqual(configStatus, this.state.configStatus)) {
             this.asAdmin(() => {
@@ -303,7 +304,7 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
 
         switch (this.internal.calibrationMode) {
             case CalibrationMode.Enabled:
-                if (!this.state.supportsCalibration) {
+                if (!this.internal.supportsCalibration) {
                     // Should never happy normally because mode attribute should never be set
                     throw new StatusResponseError("Calibration not implemented", StatusCode.Failure);
                 }
@@ -348,7 +349,7 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
         direction: MovementDirection,
         targetPercent100ths?: number,
     ) {
-        if (this.state.disableOperationalModeHandling) {
+        if (this.internal.disableOperationalModeHandling) {
             return;
         }
         switch (type) {
@@ -389,7 +390,7 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
      * completes.
      */
     #prepareMovement(type: MovementType, direction: MovementDirection, targetPercent100ths?: number): void {
-        if (this.state.supportsCalibration && this.internal.calibrationMode === CalibrationMode.Enabled) {
+        if (this.internal.supportsCalibration && this.internal.calibrationMode === CalibrationMode.Enabled) {
             return this.env.runtime.add(this.#executeCalibrationAndMove(type, direction, targetPercent100ths));
         }
         if (type === MovementType.Lift && this.state.configStatus.liftMovementReversed) {
@@ -410,7 +411,10 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
                             ? MovementDirection.Close
                             : MovementDirection.Open;
                 }
-                if (!this.state.disableOperationalModeHandling && direction !== MovementDirection.DefinedByPosition) {
+                if (
+                    !this.internal.disableOperationalModeHandling &&
+                    direction !== MovementDirection.DefinedByPosition
+                ) {
                     this.state.operationalStatus.lift =
                         direction === MovementDirection.Close
                             ? WindowCovering.MovementStatus.Closing
@@ -431,7 +435,10 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
                             ? MovementDirection.Close
                             : MovementDirection.Open;
                 }
-                if (!this.state.disableOperationalModeHandling && direction !== MovementDirection.DefinedByPosition) {
+                if (
+                    !this.internal.disableOperationalModeHandling &&
+                    direction !== MovementDirection.DefinedByPosition
+                ) {
                     this.state.operationalStatus.tilt =
                         direction === MovementDirection.Close
                             ? WindowCovering.MovementStatus.Closing
@@ -452,7 +459,7 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
 
     #executeCalibrationAndMove(type: MovementType, direction: MovementDirection, targetPercent100ths?: number) {
         let calibration;
-        if (this.internal.calibrationMode === CalibrationMode.Enabled && this.state.supportsCalibration) {
+        if (this.internal.calibrationMode === CalibrationMode.Enabled && this.internal.supportsCalibration) {
             this.internal.calibrationMode = CalibrationMode.Running;
             calibration = this.executeCalibration();
         }
@@ -471,7 +478,7 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
      * @protected
      */
     protected handleStopMovement(): MaybePromise {
-        if (this.state.disableOperationalModeHandling) {
+        if (this.internal.disableOperationalModeHandling) {
             return;
         }
         if (this.features.positionAwareLift) {
@@ -710,19 +717,14 @@ export class WindowCoveringServerLogic extends WindowCoveringServerBase {
 
 export namespace WindowCoveringServerLogic {
     export class Internal {
+        /** Does the device supports calibration? */
+        supportsCalibration: boolean = false;
+
         /** Status of the Device Calibration mode. */
         calibrationMode: CalibrationMode = CalibrationMode.Disabled;
 
         /** Status of the Device Maintenance mode. */
         inMaintenanceMode: boolean = false;
-    }
-
-    export class State extends WindowCoveringServerBase.State {
-        /** Does the device supports calibration? */
-        supportsCalibration: boolean = false;
-
-        /** Does the device supports maintenance mode? */
-        supportsMaintenanceMode: boolean = true;
 
         /**
          * Disable OperationalMode and position value management.
@@ -730,6 +732,11 @@ export namespace WindowCoveringServerLogic {
          * absolute values according to the feature set) according to the Matter specification!
          */
         disableOperationalModeHandling: boolean = false;
+    }
+
+    export class State extends WindowCoveringServerBase.State {
+        /** Does the device supports maintenance mode? */
+        supportsMaintenanceMode: boolean = true;
     }
 
     export declare const ExtensionInterface: {